Saddle Point Function
This function finds saddle points in a Matrix, a saddle point is defined as an element whose value is greater than or equal to every element in its row and less than or equal to every element in its column. Note that there may be more than one saddle point in a Matrix. Return a Matrices that has exactly two column. Each row of indices corresponds to one saddle point with the first element of the row containing the row index and the second element containing the column index. If there is no saddle point in a matrix M then the indices is an empty array.